main assignment is done with sweet alert

+++ src/Pages/List.jsx
@@ -1,6 +1,7 @@
 import {   useEffect, useState } from "react";
 import useAuth from "../hooks/useAuth";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 // import { AuthContext } from "../providers/AuthProvider";
 // import CardMyList from "../Components/CardMyList";
 
@@ -22,16 +23,36 @@ const List = () => {
     },[user, control])
 
     const handleDelete = (id) =>{
-        fetch(`http://localhost:5000/delete/${id}`, {
-            method: 'DELETE'
-        })
-        .then((res) => res.json())
-        .then(data =>{
-           if (data.deletedCount > 0) {
-            setControl(!control)
-            
-           }
-        })
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+          }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/delete/${id}`, {
+                    method: 'DELETE'
+                })
+                .then((res) => res.json())
+                .then(data =>{
+                   if (data.deletedCount > 0) {
+                    setControl(!control)
+                    Swal.fire({
+                        title: "Deleted!",
+                        text: "Your file has been deleted.",
+                        icon: "success"
+                      });
+                    
+                   }
+                })
+             
+            }
+          });
+
+       
     }
     return (
         <div className=" bg-white">
